+++ pocketbase_helpers/lib/pocketbase_helpers.dart
@@ -6,3 +6,4 @@ export 'src/base_helper.dart';
 export 'src/helper_utils.dart';
 export 'src/pocketbase_connection.dart';
 export 'src/auth_helper.dart';
+export 'src/file_helpers.dart';